requirements.txt
# ─── Web Framework ─────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Database ──────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4

# ─── Exchange / Broker ─────────────────────────────
python-binance==1.0.19
ccxt==4.3.6               # unified exchange interface

# ─── Market Data ───────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b         # technical indicators
aiohttp==3.9.5

# ─── Backtesting ───────────────────────────────────
backtesting==0.3.3

# ─── Alerts ────────────────────────────────────────
python-telegram-bot==21.3

# ─── Scheduling ────────────────────────────────────
apscheduler==3.10.4

# ─── Security ──────────────────────────────────────
python-dotenv==1.0.1
cryptography==42.0.7
python-jose[cryptography]==3.3.0

# ─── Utilities ─────────────────────────────────────
httpx==0.27.0
tenacity==8.3.0            # retry logic
structlog==24.1.0          # structured logging
python-multipart==0.0.9

# ─── Testing ───────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0              # test client

# ─── Analytics & Reporting ─────────────────────────
matplotlib==3.9.0            # chart generation (PNG)
reportlab==4.2.0             # PDF export
openpyxl==3.1.2              # Excel export (also used by pandas)
aiosmtplib==3.0.1            # async email delivery

# ─── Dev Tools ─────────────────────────────────────
black==24.4.2
ruff==0.4.4
mypy==1.10.0
